Add `rule` parameter to `parse` and enforce end of input
Certain tests require enforcing the end of input was reached. The `parse` method
is the intended place for this, but also needs a `rule` parameter so it can also
be used with statements/expressions (not just the source rule).

This approach also keeps DSLs in mind, which may want to be called with
different rules as well, such as statement/expression contexts with `#python` or
query/update statements with `#sql`. Hence, a string literal is better over an
enum for now.