Refactor test structure to use nested tests
Nested tests allow grouping beyond just parameterized tests, which is useful for
having separate test structures for multiple categories (such as literal
expressions with scalars, lists, and objects) and error conditions (including
the index/message).

To allow this, it is necessary for each class to use `Lifecycle.PER_CLASS` for
parameterized test `@MethodSource`s, as inner classes in Kotlin cannot have a
companion object which would be the alternative approach. This is set as the
default in `junit-platform.properties`.