Add AST helpers and standardize formatting in parser tests
The `statement(name)` and `expression(name)` helpers are for creating named
statements/expressions, which should be used only when the type of the AST is
not important for the test (effectively, parsing restrictions).

Additionally, tests should create the AST on a new line after the input so the
full AST is easily visible (especially since some of the test input is getting
lengthy).