Implement DSL language for extracting DSL input
This language parses the contents of a DSL into a list of string literals and
interpolated arguments between each literal (TODO). DSLs can either be on the
same line, which disallows newlines or braces in the language, or start on a
new line and use indentation to identify the end.

This still needs to be connected to the Rhovas parser.