Implement evaluation of core statements
Specifically: block, expression, declaration, assignment.

This also requires restructuring the test helpers to work with statements. The
design here is to use a `log` function which logs (and returns) the argument to
a string which is used as the result of the test. This approach can also be used
for testing execution order and potential re-evaluation of expressions (for a
later time).