Implement evaluation of loop related statements
Specifically, for/while/label/break/continue. For currently only allows lists as
an Iterable type does not exist.

The tests for these are merged into for/while, since label/break/continue is
dependent on a loop. There's probably a better way to divide this up but for now
is sufficient to avoid regressions during refactors.