Add reflect annotation for defining library functions
This is a proof-of-concept for defining functions using reflection and
annotation data to perform type checking and unwrapping. The main 'hack' is that
arguments and return values can be unwrapped/wrapped depending on whether the
method type is an environment Object, automatically performing type checking
based on the provided types (the runtime type of the value is not sufficient to
know the language type). Type checking can be bypassed using an empty type.