Implement function declarations
This is considered a statement for two reasons: first, simplicity with the
existing implementation as opposed to adding another level to the grammar; and
second, to maximize compatability in a REPL environment by allowing statements
to be the top-level construct. Rhovas source files will likely maintain a
standard class/member/code structure, but it is important to ensure the grammar
can support top-level statements for scripting environments.

Additionally, tests do not currently account for scope handling, which applies
to other ASTs as well and requires a new testing approach.