Implement evaluation of pattern matching
This is an unoptimized solution that relies on having a global evaluation state.
The main issue is to determine where (and how) varargs can be used in
destructuring, and how complex patterns should be handled. For example,
`{[key, value]*}` for applying a pattern check to object fields. Related,
consider restricting varargs to inside of destructures.