Improve lexer error messages and fix input diagnostic calculations
Input calculations didn't account for lines that do not end with newlines (even
if enforced, the diagnostic still needs to handle this case to produce the error
for it). Tests have also been added which specifies the error summary (which
should be consistent for purposes of searching) and range (for obvious reasons).
The detail message is not include in tests.