Add parser context for error diagnostics
The context manages a stack of input ranges that should be displayed in error
diagnostics. Generally speaking, this is lines that can introduce indentation
such as statements with blocks (like `if`), array/object literals, and so on. In
the case of chaining (binary/secondary expressions), the start of the chain and
the start of the previous part are both included. Once context information is
stored in the AST that should be used instead of manual calculations.

Additionally, for implementation simplicity with DSLs the context is considered
part of the lexer state (as lexer errors are parser errors) so it can be shared.

This is currently 'best-effort' as far as what's included in the context. Later
on, it may be feasible to provide better results by taking newlines into account
(ideally with restrictions on where newlines can occur).