Add AST context for error diagnostics
This allows the AST to store source information. The context is implemented as
an internal field for simplicity that way it is not required for parser tests.
The context is designed such that the first and last elements are the ranges of
the first and last tokens over that AST. In the future, more tokens could be
stored as required by the specific AST class.