Add diagnostic error messages for the Evaluator
This is the first step in the process to include a summary and details message,
the next step will be implementing context handling which requires tracking the
context and stacktrace in the evaluator as well as ensuring the necessary
information is available when evaluating functions (especially library ones).

Additionally, the main Evaluator instance has been made global for simplicity;
long-term this will have to be addressed for multithreading.