Refactor environment classes to remove runtime requirement
The current environment classes store runtime information which would not be
available during semantic analysis (in particular, variable values and method or
property receivers). This is a quick solution that uses lateinit properties,
which makes the implementation messier for now (especially tests), but this
should be improved in the process of implementing semantic analysis.