Remove semantic analysis assertions from Evaluator
This cleans up the Evaluator to remove assertions now handled by the Analyzer to
make it easier to migrate the Evaluator to use RhovasIr. For simplicity, error
tests will look for Analyzer errors first before evaluation (until these can
be properly separated).

This makes the Evaluator not defensive, as there are no checks for potential
type errors. In the future it may be worth including defensive checks for things
not enforced by static typing.