Fix analysis/evaluation handling of variables/functions
Storing environment variables/functions in RhovasIr overlooked the necessity of
dynamic dispatch for variables (for recursive functions) and methods (standard
object-oriented programming). This fix introduces unbound methods which are
stored in the RhovasIr during static analysis that can be bound at runtime.

This does not address handling local variables or functions which refer to
properties/methods in the current class; this is a challenge best handled when
classes are implemented. Furthermore, the is not much static type safety for
managing proper scope and environment classes between analysis and runtime
(manual assertion with casting) which is important for DSL analysis.

Finally, tests are passing though a few have been disabled as they require type
inference during static analysis.