Add Dynamic type and runtime type checking
This type allows type checking to pass when the actual type is unknown due to
type inference or generics (both currently unimplemented). Runtime type checking
has been re-added to the Evaluator as well.

In the future, it may be ideal to have an Evaluator specification and AST/IR
system that does not require semantic analysis to be performed. This is useful
for REPL behavior and also separation of concerns, since runtime checking needs
to be handled anyways it makes sense to have this be the baseline and then any
static analysis act as a layer of enforcement on top.