Implement environment support for subtyping and generics
The main idea is to represent each class as a Base type, which is bound with
generics through a Reference type. Each Base type stores a complete scope of
all variables/functions, which is necessary for fast lookup and avoiding the
diamond problem. Generics are bound lazily when needed for lookup or checking
inherited types, which may be expensive for checking subtypes (collecting all
supertypes may be a better option).