Add subtyping and generics to standard library types
This is built on a house of cards on a foundation of sand.

First, the standard library itself is constructed a Reflect.Type parameter used
to define types. The class is annotated with a type defining generics; it does
not support bounded generics. Long term writing the declaration in Rhovas is
likely the best option.

When type checking, the main issue is that the subtype behavior of generic types
is different at runtime vs compile-time due to type erasure (since generic 'T'
becomes assignable to it's bound instead of a separate type). The solution used
here is to avoid performing type erasure at runtime by binding all generics to
Dynamic, which is incredibly hacky.