Update push-pull context system with separate `ContextItem`s.
This is another abstraction layer that moves the handling of individual context
items to the specific analyzer, which will allow DSLs to define their own
contexts. The `scoped` method has also been replaced with `analyze`, which uses
the context instead and merges only when the context is not explicitly given.