Implement string interpolation
Lexing for string interpolation requires switching lexer modes from the parser
(since the lexer can't determine when interpolation stops). This adds a mode
variable to Lexer than can be used to change lexer modes - there's probably a
cleaner solution that can be explored when more DSL work is done. Like with
DslParser, interpolated strings are stored as a combination of literals and
arguments.