Add input context for semantic analysis and IR
Previously, context was not being tracked by the analyzer or assigned to the IR
(though the field existed). This is one step in fixing this, which just ensures
that the analyzer has basic context tracking (first token of the AST) and the IR
context is set.

Moving forward, this should ensure the analyzer uses the ideal context for each
AST (such as function names) and non-local context (such as declarations). The
specification for this needs some thought.