Change exception context from pull to push
Instead of pulling thrown exceptions, it now pushes caught exceptions. This
allows uncaught exceptions to error at the location where the exception is
thrown instead of at the end of the try/catch or function. However, this no
longer errors on unknown exceptions (which for now is preferable, these should
be handled as warnings because it's better for testing and REPL environments).