Standardize format for evaluator tests
Updates evaluator tests to use multiline strings and adds some helper functions.
These tests need a significant update as they're somewhat all over the place and
many are failing with errors in the analyzer rather than the evaluator (which
I believe was a temporary workaround when the analyzer was still being created).