Remove Evaluator tests erroring from static analysis
Most of the tests for error handling in the Evaluator are now covered by the
Analyzer, resulting in these tests never actually reaching the Evaluator. These
tests serve no purpose and can be removed for now. Later on, these types of
tests can be re-implemented such that they reach the Evaluator as intended.