Refactor function/method resolution into the Analyzer
Previously, lookups in the Analyzer used the Scope method taking a list of
types. This significantly limited the quality of errors for invalid arguments
and prevented bidirectional type inference for literals/lambdas/etc.

The code is a bit repetitive but given the differences between functions,
methods, and pipelining (even with/without qualifier) and amount of parameters
needed it's better to duplicate the structure in their respective functions.