Add nested type access
This is a partial solution to support nested types (Module.Type) within the type
rule of the grammar. However, the general case of handling qualified access for
fields/functions, proper constructor calls, and pipelining is much harder and
still needs to be done. This primarily relies on determining how specialized the
AST structure should be and whether to implement first-class types at this time.