Remove additional Java standard library uses
This is part of compatibility for non-jvm targets like Kotlin/JS. Some things
like error messages are a bit less optimal (Kotlin/JS only supports simpleName,
for instance) but is an acceptable tradeoff for now.

The only Java class still in use is BigInteger/BigDecimal, which is something
that needs a bit more thought on how to handle effectively cross-platform.