Add Analyzer support for direct defintions (resolves #9)
This is the first step which solves the main issue of having duplicate
definitions by having the Evaluator check for declarations/definitions.

The second step is to address nested classes (either within other classes as
part of structural code or within functions). As part of this, the Type system
needs to be redesigned to allow declarations for non-concrete classes.
Consequently, all runtime Object should only have concrete Types which maintains
type safety for definitions.