Add JavaScript DSL for JS environments
Not perfect, but acceptable as a proof-of-concept. The main trick here is to
use `eval?.(source)` to force an *indirect* eval, which runs in the global scope
instead of locally.

This also removes the end of input check from DslParser, since DSLs are intended
to be embedded. In the future this might be replaced with an `embedded: Boolean`
parameter to help with testing more complex DSLs.