Add iterative factorial program and fix generic subtype issue
The iterative factorial uses a normal for loop with range, which failed
typechecking due to a subtyping issue with unbound generics. This was solved
for function typing previously with bindings, so the solution here is to
remove the two separate isSubtypeOf methods (long overdue) and default the
first to an empty map. This should theoretically help with generic constraints
within the same type, though this is not an edge case currently supported.