Refactor AST/IR to include a dedicated Constructor expression
This is primarily to better support first-class types, but also opens the door
for initializer blocks as well. Constructors are stored as normal functions with
 an empty name. For simplicity with testing, constructors for Exception and
 Nullable have been added to the standard library (long overdue).

Additionally, this fixes an important Analyzer bug for function resolution where
variant types weren't accounted for when binding generics. This lookup process
is now duplicated four times and needs to be refactored anyways, so for now the
actual resolution is performed through the normal scope lookup as the single
source of truth.