Implement first-class types
First-class types (type literals) allow types to be used as expressions. This
is intended for two purposes - .is/.as/.to and macros - both of which are
handled during semantic analysis.

There are currently no plans (or specifications) for how types work at runtime.
In the future this may either be officially supported if there are compelling
use cases, or alternatively prevented by the analyzer entirely. While this could
be done by special-casing the above in the parser, the convention has always
been to delegate as many semantic validations to the analyzer as possible for
better IDE support long-term.