Add expression blocks with lambda support
This is the first step towards statements-as-expressions. For now, both
statement and expression blocks are supported (somewhat hacked via semicolon
checks) but in the future this will be condensed to expressions only with any
statement restricts handled in semantic analysis.

Exhaustive unit tests are intentionally left out as we transition to more
integration-level tests via example programs (less volatile and works with the
KotlinJS distribution).