Add is/as/to, Type<T> generic, basic inheritance, and fix invoke returns
This implements is/as/to functions, which make use of first-class types. The
Type class has also been given a generic type T to access the actual type of
first-class types statically. Furthermore, basic inheritance to automatically
include is/as/to functions in subtypes has been added as well - this does not
support overriding, hence relies on final methods / concrete classes.

Additionally, this also fixes two issues with computing function/method returns.
First, cascade should not wrap nullable return types with another layer of
nullable - they should be merged by definition. Second, methods should pass the
return type of the final method object (and not the backing function) in case
there are any discrepancies (such as specialization) - it's not clear if this is
a current issue, but is conceptually incorrect so worth changing.