Implement Struct methods and redesign analyzer phases
To handle declaration/definition phases, a new intermediate IR for storing type
information for future use has been added. The general idea is that these phases
can be done beforehand (for structural code) or on demand (for executable code),
but it's unclear how well this will scale.

Also, this currently fails Kotlin/JS tests - there appear to be multiple issues
with the generated IR causing failures that are near impossible to debug. This
will likely result in moving more tests from the JVM level for analysis/parsing
into common to allow testing with Kotlin/JS as well.