Change methods to require an explicit 'this' parameter
Instead of automatically adding a 'this' parameter to all functions within a
component, require an explicit 'this' parameter. For simplicity, this is
implemented by defaulting the type of the first argument to the current
component if it is not explicitly given. In the future this will likely require
the parameter name to be 'this'.