Add basic initializer functions for structs
This allows structs to define custom initializers using `init`. Due to other
limitations, initializers cannot assign to final fields (need support for split
declaration/initialization for variables and tracking of struct fields (not
properties) for initialization. Long-term this may not be needed due to
initializer blocks, but would be helpful for other variables nonetheless.

Furthermore, custom initializers do not currently override the default struct
initializer. This is likely to remain the case due to the nature of structs, and
additional initializers are for convenience.