Add initialization context for validating uninitialized variables
This checks that variables are initialized before used, are only initialized
once (if immutable), and not partially initialized across branches. This is a
step towards allowing struct initializers to initialize immutable fields.

This also comments out a runtime check in the Evaluator for assignment to
immutable variables, which would require rebuilding a large part of the context
in the analyzer to enforce. There is a general question here of how much the
Evalator should enforce behavior at runtime; it may be useful to treat it as a
fully dynamic system to keep the implementation simple, avoid duplication with
semantic analysis, and make it easier to test the Evaluator itself.