Implement initializer statements
This implements initializer statements, which are used in initializers to
atomically initialize fields. This is primarily planning for inheritance
support - superclasses in Java can't safely call abstract/virtual methods in the
constructor since child fields haven't been initialized yet, but this approach
resolves this limitation since fields will be initialized as part of the `super`
call.

The current implementation is missing static validation of fields in the
initializer as there currently isn't infrastructure to access fields of the
current struct. Having struct types would be helpful too.