Remove defaults for function parameter types
Function parameters must have a defined type (with the exception of a `this`
parameter for methods) instead of defaulting to `Dynamic`. The return type,
however, will default to `Void` as before.

Note that Dynamic is still used internally in places where type inference isn't
sophisticated enough yet (e.g. lambdas).