Implement type definition for Struct types (#14)
This is the first step for implement Struct types, which handles the core type
definition. Note that there is both a Struct type representing the class Struct,
and a Struct Type class for managing fields and subtype rules.

The intended behavior of Struct types and structural subtyping overall is still
undecided, but for now this should follow standard structural subtyping rules.