Refactor toType methods into to(Type) variant
Following the addition of `to`, methods like `toString` should internally be
represented using the `to(String)` form. This also includes other conversion
methods like `String.to(Atom)`.

Additionally, this adds a helper function for invoking `to(String)` on an
`Object` via `object.methods.toString()` - it's a little hacky, but works well
and this method is used frequently enough to justify it.