Implement Result types, coalescing, and fix related errors (#2)
This adds Result types, which representing a success/failure type for error
handling. This also update internal builtins for coalescing to using Results.
Furthermore, the approach for null literals has changed to represent null as
being a Nullable<Dynamic> type, which works naturally for subtyping rules and
is the same idea as Kotlin's approach.

Additionally, this corrects a number of related errors for types/inheritance:
 - Type.Generic.isSubtypeOf only bound generics if the supertype was Generic.
 - TypeInitializers must be listed in initialization order (for inheritance)
 - Inheritance method identification should have used isSupertypeOf
 - Inheritance needs to bind functions using the supertype generics
 - Inheritance needs to define functions using the defined name (for operators)