Implement analyzer context for bidirectional type inference (Resolves #13)
This is a starting point for bidrectional type inference, which allows the type
of an expression to be inferred based on surrounding context (e.g. List/Tuple
literals or Lambda parameters). The current plan is to allow this for literals,
but avoid any implicit conversions/coercions for other expressions.

This also fixes some related issues, most notably with subtyping:
 - Fixes variant bound being updated for non-subtypes (.takeIf { true })
 - Fixes cases where generic bindings were not used/passed in subtype checks