Improve subtype checking for generics with Tuple/Struct types (#14)
This removes duplication of generic subtype logic in Tuple/Struct types thorugh
the use of TYPE, which forces a subtype check into the same generic logic that
handles variance. This logic has also been fixed for handling nested generics
or variants that arose out of lambda arguments (namely, when the subtype
contains generics/variants reversing the subtype check does not work for
invariance; for now checking the base types are equal seems to work fine).