Add Struct inheritance and constructor typechecking
This changes structs to inherit from the Struct type (includes to(String)) and
change the default initialize type to the appropriate Struct (excludes any
fields with default values, since there's currently no way to represent
optional struct fields).

Type.Base and Library have been updated to work with user-side inheritance,
which needs to work with late initialization and be able to use the inherit
method. Long-term, it may be worth enforcing these initialization phases and
other assumptions related to the structure of types.