Add helper for equality checks and simplify equality operator logc
Adds an Object.methods.equals helper (similar to toString) that can be used for
equality checks. The logic for binary equals now uses this helper over storing
a method and relying on manual subtype checks to see if it is invokable.

Note that the evaluator no longer includes a stacktrack around equals calls in
the same was as toString and other manually-invoked methods. The current
stacktrace approach needs a significant redesign anyways (as does runtime error
handling in general).