Implement bang annotations for Result/Exception isomorphism
This implements bang annotations, which allow functions that return Result
value to throw Exceptions (and in turn, lack thereof allows functions that
throw Exceptions to return Results). This makes chaining operations and working
with errors in a functional style very natural.

For simplicity, this modifies the lexer rule for identifiers to include the bang
bang suffix (otherwise it would need to take into account whitespace to
distinguish name! == ... from name !== ...). The analyzer doesn't enforce name
requirements so this can result in some oddities, but this is acceptable for now
and can be addressed in the parser later on.