Replace Object usages with typed Structs
This allows Struct instances to exist directly (from object literals, which may
eventually be renamed) and used in destructuring. Object literals infer specific
fields (for which mutable object literals need to be considered later), while
destructuring simply uses a Dynamic struct (to be addressed later).

This also fixes a typo with extracting tuple arguments for lambdas.