Replace Object with a general-purpose Map type (#2)
The Object type was originally design for object literals which required Atom
keys, which is now obsolete following Structs. The new Map type supports
arbitrary keys, along with the basic utility functions for maps. Furthermore,
object literals can be inferred to Maps.

This also patches a type issue with subtyping and binding two generic types
(binding a generic to itself results in an infinite loop). A proper solution
needs to account for the source of generic bindings, not just names, akin to
type variable IDs.