Implement Equatable/Comparable interfaces (#2)
These interfaces are used to identify types that may be used with ==/<=>
operators. The primary limitation is that they cannot support generic predicates
(e.g. List<T> is Equatable if and only if T is Equatable), particularly for
Comparable (since non-Equatable types just return false). This is a complex
design area and the type system isn't sufficient to support this completely.

Additionally, there are default implementations of these which should work for
most built-in types (like toString).