Implement Hashable type, fix type inheritance binding (#2)
Adds a Hashable type for objects that can be used in hash lookups (which
naturally must also be Equatable). This is currently only used for Map keys, but
is being added in preparation for Sets.

Also fixes a bug with type inheritance checking for overloads before binding
generics rather than after.