Setup Kotest for JS test support
The current unit tests are JVM only because they use JUnit for parameterized
tests, which is not offered by kotlin.test. Kotest is a Kotlin testing library
that can support this, which in turns allows future tests to work with JS. The
existing tests will be migrated in a few steps.

The tests use StringSpec in part to avoid overlapping with FunSpec.test. As more
tests are migrated, this limitation will likely be addressed in the process of
supporting context for JVM without breaking JS (JS targets do not allow context
to be used due to limitations of nesting with promises and suspend functions).