Support test suites on Kotlin/JS, start migrating JVM tests
This allows test suites to be used with Kotlin/JS through RhovasSpec (see class
docs for details). TLDR; to work around a limitation with nested promises this
registers all tests synchronously and registers nested tests as top-level tests
with a prefix when running with Kotlin/JS.

Additionally, the existing tests for full-programs has been moved into a
programs package (matching file location) and the DslLexerTests have been
migrated to commonTest as a proof-of-concept.