Migrate RhovasAnalyzer tests to commonTest
Continuation of Kotest migration for Kotlin/JS support. As previously, this has
some restructuring but the test content is nearly identical (exceptions being
with how necessary variables/functions/types are created, largely through using
new additions to the standard library).

The analyzer tests in particular have a lot of room for improvement to be done
in a later commit. The largest impact would probably be additional helpers for
building expected IRs and related simplifications. Tests that properly check
error messages would also be useful.