Migrate Evaluator tests to commonTest, fix skip/focus for Kotlin/JS
Continuation of Kotest migration for Kotlin/JS support. As previously, this has
some restructuring but the test content is nearly identical (exceptions being
with how necessary variables/functions/types are created, largely through using
new additions to the standard library).

This also fixes the use of skip/focus tags for Kotlin/JS, as these tags need to
be moved to the start of the name to be picked up properly since the actual name
is appended onto the prefix for the suite.