Add support for null coalescing with index access
This isn't standard in many languages, but there's no real reason to not support
it and it's often better than having to switch to named functions. Note that
coalescing with index assignment is not currently supported, as that's a little
"too magic" and disguises the fact that the value would not be evaluated.