Implement positional Struct constructors, fix object literal inference (Resolves #21)
A positional constructor allows initializing structs by passing all fields in
definition order, thus avoiding the need to explicitly specify field names. All
fields are required to be given values as default arguments are not implemented.

This also fixes an issue with object literals using the mutability of the
inferred type, which leads to inconsistent behavior. For now, all object
literals are assumed to have mutable fields. Additionally, the toString output
for Type.Struct has been updated to include this to assist future debugging.