Allow ensure after return within a block, fix block expression scope (Resolves #23)
Ensure statements can be defined after a return to allow easier access to the
return value (val). The implementation here relies on this occurring within a
block and only applies to return statements, which may not be ideal but is
reasonable for now to get the intended logic present.

Also fixes block expressions not being analyzed/evaluated in a new scope.