Improve analyzer test coverage, fix discovered bugs
This adds tests for untested AST/IR classes as well as edge cases for other
conditions that were not tested (e.g. error messages). There are still a few
places that aren't covered, but this handles the bulk of remaining work.

Additionally, this fixes multiple issues in the analyzer discovered by the new
tests (mostly related to type inference).