Implement element access for Tuples (#26)
Tuple elements can now be accessed like other properties like `.0`, which allows
the element type to be statically known. This reuses the existing Property AST
as the backend behavior is identical and it's not currently worth special-casing
this. Additionally, Tuple literals have been changed to be mutable by default
(same as Structs).