Fix lambda invocation arguments and exception analysis (Resolves #24)
This fixes how lambda arguments are handled through the invoke method and allows
a Dynamic exception type to bypass analysis in the case when no exceptions are
declared. Tests for this will be added to the Evaluator as part of improving
test coverage, but for now this should allow lambdas to work correctly in many
more use cases.