Implement user-defined operator overloading (Resolves #22)
This adds support for the `op+` syntax in function declarations used to specify
operator overloading. Currently, only the basic math (+, -, *, /) and indexing
([], []=) operators are allowed for overriding. Furthermore, there are no checks
for whether the signatures of these functions are valid - it is likely that
operators will need to be added into Function.Declaration for this, particularly
with overriding once inheritance is added.