Improve handling of generics, particularly for Tuple/Struct types
This is a minor improvement to how generic types are handled in the system,
particularly when dealing with reference declarations. The ANY accessor has now
been replaced by GENERIC, which returns the appropriate declaration for that
type rather than using DYNAMIC for all generics. This hack was originally in
place for subtype checking, which now explicitly uses DYNAMIC (this should not
need to use DYNAMIC, but that's an issue for later on).

Additionally, helper methods for Tuple/Struct types have been added along with
an assertion during initialization to help enforce proper type construction. It
would be helpful to add more assertions, but for various reasons (initialization
order and recursive construction during subtype checks) this will have to wait.