Implement proper algorithm for unification (Resolves #20)
This adds a new Type.unify method which performs unification (if that's even the
correct term) to combine types during type inference. There are likely issues
with how generics are handled, but at the very least it improves type inference
for Maps (and doesn't break anything else) so at least is a step in the right
direction.