Fix Kotlin/JS tests being disabled due to focus, cleanup debugging
Kotest only supports focus for top-level tests due to mixed registration and
execution limitations. On JVM, the focus tag was thus effectively a no-op and
thus tests worked as expected. However, on JS we restructure nested tests to be
all top-level (also for working around a limitation with nested promises) which
effectively disables all other test cases.

For now we can work around this, but eventually RhovasSpec may get updated to
incorporate another workaround to allow focus to be used on JVM.