Fix analyzer context for components/functions, improve context management
When entering a component/function, the context needs to be set to fresh values
that are accurate for a new executable scope (e.g. initialization). This also
makes some minor QOL changes, the largest being inference context is now
non-nullable as Type.ANY (or similar) should have the same effect.