Standardize lambda invocation to use Tuples, not Structs
The original lambda invocation logic use named arguments to create a struct for
the default val parameter. These names do not work with Tuple types (as Tuples
are not intended to be named). The proper solution to support this is ordered
Structs, but rather than maintain two calling conventions in the interim this
removes the Struct approach in favor of Tuples.