Force type erasure for generic types at runtime
Runtime types in the evaluator cannot directly use static types due to the
potential of unbound generics which can be leaked deeply into the object
structure (which affects runtime typechecking).

Ideally, the long-term solution is to support as much runtime type checking as
reasonable. It is unclear if full reified generics are necessary for this, or if
there's an alternative solution that preserves a sufficient amount of runtime
type information. This may be another use case for an invocation context that
would also contain things like the evaluator and debug information.