Improve type test coverage, add new delegate helpers for generics (#19)
Final part of test coverage work for the type system. Tests are designed to
cover the complex, interconnected edge cases necessary, hence coverage itself is
less meaningful compared to previous parts. Additionally, some delegate helpers
for constructing generic types have been added to improve testing, readability,
and minimize broken invariants.

This fixes a few bugs in the type system. First, Tuple/Struct types should now
properly handle mutability with subtyping (namely, a mutable field is not a
subtype of an immutable one). Unbound variants should now work properly with
Dynamic. Finally, there is a newly discovered (but not yet resolved) issue with
relation to the order variant types are bound in with Dynamic when generics are
used multiple times in a functions (issue to come).