Implement core inheritance and initializer support
This is a first pass at inheritance which should cover the core functionality
(inheritance of fields/methods and constructors). The main limitations right now
are with restrictions (e.g. interfaces, virtual/abstract classes and methods,
etc.) and initializers (quirks with initializing fields and maintaining proper
subtyping in the parent class).