Add inheritance modifiers, implement inherits type validation
This adds modifiers for inheritance (virtual/abstract) that are used with
components/methods for inheritance. This covers syntax support (which is helpful
for writing accurate programs even if the validation comes later) and validates
the inherited component permits inheritance. For now, interfaces are considered
abstract classes (approach TBD once interfaces are added).

There is no validation for methods (e.g. override). The current approach for
inheriting methods doesn't work with this; as it effectively inherits any parent
methods that are undefined. Instead, a different approach that handles
inheritance first followed by method definition needs to be done (or a fully
separate validation approach).

Additionally, note the require precondition added to Type - my hope is that
more of these kinds of preconditions can be explicitly specified moving forward.
This is mainly limited by hacks from standard library initialization.